+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add type annotations as a way to declare schema
 - Deprecate "free" queries
 - Create simpler developer contract for databackend
+- Snowflake native data-backend implementation
 
 #### Bug Fixes